# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Agendamento API",
        "description": "API REST para gerenciamento de agendamentos com autenticação JWT, gestão de dias fechados e QR Code para validação. Esta API permite criar, listar, visualizar e cancelar agendamentos, além de gerenciar dias bloqueados e validar QR Codes.",
        "termsOfService": "http://localhost:8000/terms",
        "contact": {
            "name": "Equipe de Desenvolvimento",
            "url": "https://github.com/seu-usuario/agendamento",
            "email": "dev@agendamento.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/admin/gestao-dias": {
            "get": {
                "tags": [
                    "Gestão de Dias"
                ],
                "summary": "Listar dias bloqueados",
                "description": "Lista todos os dias com bloqueios ou restrições",
                "operationId": "a438b513ea271e1e636e22354ec3dc92",
                "parameters": [
                    {
                        "name": "data",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-15"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de dias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "data": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-12-15"
                                                    },
                                                    "tipo": {
                                                        "type": "string",
                                                        "example": "bloqueio_total"
                                                    },
                                                    "observacao": {
                                                        "type": "string",
                                                        "example": "Feriado nacional"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gestao-dias/{id}": {
            "get": {
                "tags": [
                    "Gestão de Dias"
                ],
                "summary": "Buscar dia por ID",
                "description": "Busca um dia específico por ID",
                "operationId": "86a0b39b8d25a49fa46a6043649a59bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do dia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-12-15"
                                        },
                                        "horario_inicio": {
                                            "type": "string",
                                            "example": "08:00"
                                        },
                                        "horario_fim": {
                                            "type": "string",
                                            "example": "18:00"
                                        },
                                        "tipo": {
                                            "type": "string",
                                            "example": "bloqueio_total"
                                        },
                                        "observacao": {
                                            "type": "string",
                                            "example": "Feriado nacional"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dia não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dia não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Gestão de Dias"
                ],
                "summary": "Atualizar dia",
                "description": "Atualiza um registro de dia existente",
                "operationId": "a5bc36500a1b349c08315c97b3a00591",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-15"
                                    },
                                    "horario_inicio": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "horario_fim": {
                                        "type": "string",
                                        "example": "18:00"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "bloqueio_total",
                                            "bloqueio_parcial",
                                            "manutencao",
                                            "feriado"
                                        ],
                                        "example": "bloqueio_total"
                                    },
                                    "observacao": {
                                        "type": "string",
                                        "example": "Feriado nacional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dia atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dia atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dia não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dia não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gestão de Dias"
                ],
                "summary": "Excluir dia",
                "description": "Exclui um registro de dia",
                "operationId": "3a15a2b940be93d50bd5a56feb22ee0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dia excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dia excluído com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dia não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dia não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gestao-dias/store": {
            "post": {
                "tags": [
                    "Gestão de Dias"
                ],
                "summary": "Criar novo dia bloqueado",
                "description": "Cria um novo registro de dia com bloqueio ou restrição",
                "operationId": "04ab9738f22ef8d4d489a0ba761be56c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data",
                                    "horario_inicio",
                                    "horario_fim",
                                    "tipo"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-15"
                                    },
                                    "horario_inicio": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "horario_fim": {
                                        "type": "string",
                                        "example": "18:00"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "bloqueio_total",
                                            "bloqueio_parcial",
                                            "manutencao",
                                            "feriado"
                                        ],
                                        "example": "bloqueio_total"
                                    },
                                    "observacao": {
                                        "type": "string",
                                        "example": "Feriado nacional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dia criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-12-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Digite o token Bearer no formato: Bearer {seu-token-jwt}"
            }
        }
    },
...
}